requirements.txt
# 🏇 Betfair Lay Agent Dependencies

# --- Betfair API & Streaming ---
# Core API wrapper for interaction
betfairlightweight>=2.12.0
# Streaming engine for automation and millisecond-latency updates
flumine>=1.36.0

# --- AI & LLM ---
# Google Gemini 2.5 Flash SDK
google-generativeai>=0.8.3

# --- Integration ---
# Model Context Protocol (Required for Claude Desktop integration)
mcp>=1.2.0

# --- Data Science & Utilities ---
# Required for downloading/processing Kash & Iggy ratings CSVs
pandas>=2.0.0
# Standard HTTP requests
requests>=2.31.0
# Loading .env files
python-dotenv>=1.0.0